Add support for prerender visibility state via the AMP Runtime. This allows users to physically display the AMPViewer on screen but not have the viewer change to the "Visible" state.
This is useful for cases where the AMP viewer is embedded inside another view controller and those view controllers are being added to a view hierarchy but are not yet visible.

For instance, adding the AMP viewer inside a collection view cell where the cell is not on screen yet. This allows prefetching without improperly setting visibility in those cases.